+++ pkg/cache/groupcache.go
@@ -7,11 +7,13 @@ import (
 	"context"
 	"crypto/tls"
 	"encoding/json"
+	"fmt"
 	"io/ioutil"
 	"net"
 	"net/http"
 	"path/filepath"
 	"strconv"
+	"strings"
 	"time"
 
 	"github.com/go-kit/log"
@@ -76,6 +78,18 @@ func parseGroupcacheConfig(conf []byte) (GroupcacheConfig, error) {
 
 	if len(config.Peers) == 0 {
 		config.Peers = append(config.Peers, config.SelfURL)
+	}
+
+	for i, peer := range config.Peers {
+		// Workaround for https://github.com/thanos-community/galaxycache/blob/master/http/http.go#L205-L210.
+		// If the peer has a slash at the end then the router redirects
+		// and then the request fails.
+		if strings.HasSuffix(peer, "/") {
+			return GroupcacheConfig{}, fmt.Errorf("peer %d must not have a trailing slash (%s)", i, peer)
+		}
+	}
+	if strings.HasSuffix(config.SelfURL, "/") {
+		return GroupcacheConfig{}, fmt.Errorf("self URL %s must not have a trailing slash", config.SelfURL)
 	}
 
 	return config, nil