+++ pkg/ui/react-app/src/types/types.ts
@@ -10,7 +10,7 @@ export interface QueryParams {
   resolution: number;
 }
 
-export interface Rule {
+export type Rule = {
   alerts: Alert[];
   annotations: Record<string, string>;
   duration: number;
@@ -23,4 +23,4 @@ export interface Rule {
   query: string;
   state: RuleState;
   type: string;
-}
+};