+++ test/e2e/exemplars_api_test.go
@@ -45,6 +45,7 @@ func TestExemplarsAPI_Fanout(t *testing.T) {
 		defaultPromConfig("ha", 0, "", "", "localhost:9090", qUnitiated.InternalEndpoint("http")),
 		"",
 		e2ethanos.DefaultPrometheusImage(),
+		"",
 		e2ethanos.FeatureExemplarStorage,
 	)
 	testutil.Ok(t, err)
@@ -54,6 +55,7 @@ func TestExemplarsAPI_Fanout(t *testing.T) {
 		defaultPromConfig("ha", 1, "", "", "localhost:9090", qUnitiated.InternalEndpoint("http")),
 		"",
 		e2ethanos.DefaultPrometheusImage(),
+		"",
 		e2ethanos.FeatureExemplarStorage,
 	)
 	testutil.Ok(t, err)