+++ test/e2e/metadata_api_test.go
@@ -32,7 +32,7 @@ func TestMetadataAPI_Fanout(t *testing.T) {
 		"prom1",
 		defaultPromConfig("ha", 0, "", "", "localhost:9090", "sidecar-prom1:8080"),
 		"",
-		e2ethanos.DefaultPrometheusImage(),
+		e2ethanos.DefaultPrometheusImage(), "",
 	)
 	testutil.Ok(t, err)
 
@@ -41,7 +41,7 @@ func TestMetadataAPI_Fanout(t *testing.T) {
 		"prom2",
 		defaultPromConfig("ha", 1, "", "", "localhost:9090", "sidecar-prom2:8080"),
 		"",
-		e2ethanos.DefaultPrometheusImage(),
+		e2ethanos.DefaultPrometheusImage(), "",
 	)
 	testutil.Ok(t, err)
 	testutil.Ok(t, e2e.StartAndWaitReady(prom1, sidecar1, prom2, sidecar2))